Fast-integration tests oauth tokens handling refactoring (#11440)
* Fast-integration tests oauth tokens handling refactoring

* wip

* wip

* wip
